Add search, fix stuff
Add a search command, runing `rg` in the current working directory with
the currently selected text or the identifer under the cursor.

Fix keybinding only working in the first file opened with vim.

Adjust names of buffer variables, so their origin is more explicit.

Fix display of the command in the script when featuring  escapable
characters.

Add more information to the README